requirements.txt
# ══════════════════════════════════════════════
# AI Müzik & Ses Prodüksiyon – Bağımlılıklar
# Kurulum: pip install -r requirements.txt
# GPU için: pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu118
# ══════════════════════════════════════════════

# ──────────────────────────────────────────────
# TEMEL FRAMEWORK
# ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1
pydantic==2.7.1
python-dotenv==1.0.1
streamlit==1.35.0

# ──────────────────────────────────────────────
# PYTORCH (GPU – CUDA 11.8)
# Not: CPU için: torch torchaudio (--index-url kaldır)
# ──────────────────────────────────────────────
torch==2.3.0
torchaudio==2.3.0

# ──────────────────────────────────────────────
# TTS – TÜRKÇE SES SENTEZİ
# XTTS v2 (Coqui TTS) — Ses Klonlama
# ──────────────────────────────────────────────
TTS==0.22.0

# ──────────────────────────────────────────────
# AI MÜZİK ÜRETİMİ
# Meta MusicGen (audiocraft) veya transformers
# ──────────────────────────────────────────────
audiocraft==1.3.0
transformers==4.41.1
accelerate==0.30.1

# ──────────────────────────────────────────────
# SES İŞLEME
# ──────────────────────────────────────────────
pydub==0.25.1
librosa==0.10.2
soundfile==0.12.1
scipy==1.13.0
numpy==1.26.4

# ──────────────────────────────────────────────
# METİN ANALİZİ
# ──────────────────────────────────────────────
# (Gelişmiş duygu analizi için – isteğe bağlı)
# transformers zaten yukarıda eklenmiştir

# ──────────────────────────────────────────────
# VERİ İŞLEME & GÖRSELLEŞTİRME
# ──────────────────────────────────────────────
pandas==2.2.2
matplotlib==3.9.0
plotly==5.22.0

# ──────────────────────────────────────────────
# DOSYA & DEPOLAMA
# ──────────────────────────────────────────────
pathlib2==2.3.7.post1; python_version < "3.4"
aiofiles==23.2.1
# Cloud depolama (isteğe bağlı):
# boto3==1.34.0        # AWS S3
# google-cloud-storage==2.16.0  # GCP

# ──────────────────────────────────────────────
# GELİŞTİRME & TEST
# ──────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0         # FastAPI test istemcisi
black==24.4.2         # Kod formatlama
